+++ examples/demo/features/support/env.rb
@@ -1,20 +1,12 @@
 # encoding: utf-8	You should see a paragraph character: §
 # File: env.rb
-# Purpose: Allow Cucumber to load the sample application configuration and hooks.
-# It also demonstrate what to do in your env.rb file to use the Macros4Cuke gem.
 
-begin
-  require 'simplecov'
-rescue LoadError
-  # Gobble silently the exception if simplecov isn't installed.
-end
 
 module DemoMacros4Cuke  # Use the module as a namespace
 
  
 # Class created just for testing and demonstration purposes.
 # Its instance, will record the output emitted by the steps.
-
 class TracingWorld
   # Will contain the text emitted by the steps
   attr_reader(:trace_steps)
@@ -27,9 +19,10 @@ class TracingWorld
   
 public
   # Write the given text to the error console
-  
   def show(someText)
-    $stderr.puts(someText)
+    # Replace every \" sequence by genuine "
+    unescaped = someText.gsub(/\\"/, '"')
+    $stderr.puts(unescaped)
   end
 
   
@@ -40,7 +33,6 @@ end # module
 # For testing purpose we override the default Cucumber behaviour
 # making our world object an instance of the TracingWorld class
 World { DemoMacros4Cuke::TracingWorld.new }
-
 
 
 # End of file