+++ features/step_definitions/demo_steps.rb
@@ -20,23 +20,4 @@ Then(/^I expect the following step trace:$/) do |step_text|
 end
 
 
-When(/^I leave (.+)$/) do |city|
-  show "I leave #{city}"
-end
-
-
-When(/^I visit (.+)$/) do |city|
- show "I visit #{city}"
-end
-
-
-When(/^I arrive in (.+)$/) do |city|
-  show "I arrive in #{city}"
-end
-
-When(/^I type \"([^"]*)\"$/) do |text|
-  show text
-end
-
-
 # End of file