+++ lib/macro_steps.rb
@@ -14,10 +14,12 @@
 #  And I click "Submit"
 #  """
 # The regexp has two capturing group: one for the phrase, a second for the terminating colon (:)
-Given(/^I define the step "(?:Given|When|Then) I \[((?:[^\\\]]|\\.)+)\](:?)" to mean:$/) do |macro_phrase, colon_capture, template|
+Given(/^I define the step "(?:Given|When|Then|\*) I \[((?:[^\\\]]|\\.)+)\](:?)" to mean:$/) do |macro_phrase, colon_capture, template|
   use_table = (colon_capture == ':')
   add_macro(macro_phrase, template, use_table)
 end
+
+
 
 # This step is used to invoke a simple macro-step
 # Example:
@@ -43,7 +45,6 @@ When(/^I \[([^\]]+)\]:$/) do |macro_phrase, table_argument|
   # The second argument consists of an array with couples of the kind: [argument name, actual value]
   invoke_macro(macro_phrase, table_argument.raw)
 end
-
 
 
 # End of file