+++ lib/macros4cuke/constants.rb
@@ -3,7 +3,7 @@
 
 module Macros4Cuke # Module used as a namespace
 	# The version number of the gem.
-	Version = '0.3.00'
+	Version = '0.3.01'
 
   # Brief description of the gem.
   Description = "Macros for Cucumber"