+++ spec/macros4cuke/templating/section_spec.rb
@@ -8,21 +8,83 @@ module Macros4Cuke
 
 module Templating # Open this namespace to get rid of module qualifier prefixes
 
+# Spec for abstract class Section.
+describe Section do
+  # Default instantiation rule
+  subject { Section.new('foobar') } 
+  
+  # Return a list of possible child elements
+  let(:sample_children) do
+    [ StaticText.new("Hello "),
+      Placeholder.new("user"),
+      EOLine.new
+    ]
+  end 
+
+  context "Creation and initialization" do
+ 
+    it "should be created with a variable name" do
+      lambda { Section.new('foobar') }.should_not raise_error 
+    end
+    
+    it "should know the name of its variable" do 
+      subject.name.should == 'foobar'
+    end
+
+    it "should have no child at start" do
+      subject.should have(0).children
+    end    
+  
+  end # context
+  
+  context "Provided services:" do  
+    it "should add child element(s)" do
+      sample_children.each do |a_child|
+        subject.add_child(a_child)
+      end
+
+      # Control that the addition work as expected
+      subject.children.should == sample_children
+    end
+    
+    it "should know the name all child placeholders" do
+      # Case: simple flat list of children
+      sample_children.each { |a_child| subject.add_child(a_child) }
+      subject.variables.should == [ 'user' ]
+      
+      # Case: at least one child is a group    
+      parent = Section.new('son')
+      [ 
+        subject,
+        StaticText.new("Bye "),
+        Placeholder.new("firstname"),
+        EOLine.new
+      ].each { |a_child| parent.add_child(a_child) } 
+
+      parent.variables.should == [ 'user', 'firstname']
+      
+    end
+    
+    
+    it "should expect that its subclasses render the children" do
+      error_message = "Method Section::render must be implemented in subclass(es)."
+      lambda {subject.send(:render, Object.new, {}) }.should raise_error(NotImplementedError, error_message)
+    end
+    
+  end # context
+
+end # describe
+
+
+
 
 describe ConditionalSection do
 
-  context "Creation and initialization" do
+  context "Creation and initialization:" do
 
     it "should be created with a variable name and a boolean" do
       lambda { ConditionalSection.new('foobar', false) }.should_not raise_error
       lambda { ConditionalSection.new('foobar', true) }.should_not raise_error
-    end
-
-    it "should know the name of its variable" do
-      [false, true].each do |existence|
-        instance = ConditionalSection.new('foobar', existence)
-        instance.name.should == 'foobar'
-      end
     end
 
     it "should know whether the rendition on existence of actual value or not" do
@@ -32,16 +94,9 @@ describe ConditionalSection do
       end
     end
 
-    it "should have no child at start" do
-      [false, true].each do |existence|
-        instance = ConditionalSection.new('foobar', existence)
-        instance.should have(0).children
-      end
-    end
-
   end # context
 
-  context "Provided services" do
+  context "Provided services:" do
     # Return a list of possible child elements
     let(:sample_children) do
       [ StaticText.new("Hello "),
@@ -52,15 +107,9 @@ describe ConditionalSection do
 
     # Default instantiation rule
     subject { ConditionalSection.new('foobar', true) }
-
-    it "should add child element(s)" do
-      sample_children.each do |a_child|
-        subject.add_child(a_child)
-      end
-
-      # Control that the addition worek as expected
-      subject.children.should == sample_children
-
+    
+    it "should know its original source text" do
+      subject.to_s.should == '<?foobar>'
     end
 
     it "should render its children when conditions are met" do